+++ common/pom.xml
@@ -112,11 +112,6 @@
           <version>${mariadb.version}</version>
           <scope>test</scope>
       </dependency>
-      <dependency>
-          <groupId>org.springframework</groupId>
-          <artifactId>spring-core</artifactId>
-          <version>5.1.6.RELEASE</version>
-      </dependency>
   </dependencies>
 
   <build>
